+++ pyogrio/_io.pyx
@@ -50,11 +50,11 @@ log = logging.getLogger(__name__)
 # (index in array is the integer field type)
 FIELD_TYPES = [
     "int32",           # OFTInteger, Simple 32bit integer
-    None,              # OFTIntegerList, List of 32bit integers, not supported
+    "list(int32)",     # OFTIntegerList, List of 32bit integers
     "float64",         # OFTReal, Double Precision floating point
-    None,              # OFTRealList, List of doubles, not supported
+    "list(float64)",   # OFTRealList, List of doubles
     "object",          # OFTString, String of UTF-8 chars
-    None,              # OFTStringList, Array of strings, not supported
+    "list(str)",       # OFTStringList, Array of strings
     None,              # OFTWideString, deprecated, not supported
     None,              # OFTWideStringList, deprecated, not supported
     "object",          # OFTBinary, Raw Binary data
@@ -62,14 +62,43 @@ FIELD_TYPES = [
     None,              # OFTTime, Time, NOTE: not directly supported in numpy
     "datetime64[ms]",  # OFTDateTime, Date and Time
     "int64",           # OFTInteger64, Single 64bit integer
-    None               # OFTInteger64List, List of 64bit integers, not supported
+    "list(int64)"      # OFTInteger64List, List of 64bit integers, not supported
 ]
+
+# Mapping of OGR integer field types to OGR type names
+# (index in array is the integer field type)
+FIELD_TYPE_NAMES = {
+    OFTInteger: "OFTInteger",                # Simple 32bit integer
+    OFTIntegerList: "OFTIntegerList",        # List of 32bit integers, not supported
+    OFTReal: "OFTReal",                      # Double Precision floating point
+    OFTRealList: "OFTRealList",              # List of doubles, not supported
+    OFTString: "OFTString",                  # String of UTF-8 chars
+    OFTStringList: "OFTStringList",          # Array of strings, not supported
+    OFTWideString: "OFTWideString",          # deprecated, not supported
+    OFTWideStringList: "OFTWideStringList",  # deprecated, not supported
+    OFTBinary: "OFTBinary",                  # Raw Binary data
+    OFTDate: "OFTDate",                      # Date
+    OFTTime: "OFTTime",                      # Time: not directly supported in numpy
+    OFTDateTime: "OFTDateTime",              # Date and Time
+    OFTInteger64: "OFTInteger64",            # Single 64bit integer
+    OFTInteger64List: "OFTInteger64List",    # List of 64bit integers, not supported
+}
 
 FIELD_SUBTYPES = {
     OFSTNone: None,           # No subtype
     OFSTBoolean: "bool",      # Boolean integer
     OFSTInt16: "int16",       # Signed 16-bit integer
     OFSTFloat32: "float32",   # Single precision (32 bit) floating point
+}
+
+FIELD_SUBTYPE_NAMES = {
+    OFSTNone: "OFSTNone",             # No subtype
+    OFSTBoolean: "OFSTBoolean",       # Boolean integer
+    OFSTInt16: "OFSTInt16",           # Signed 16-bit integer
+    OFSTFloat32: "OFSTFloat32",       # Single precision (32 bit) floating point
+    OFSTJSON: "OFSTJSON",
+    OFSTUUID: "OFSTUUID",
+    OFSTMaxSubType: "OFSTMaxSubType",
 }
 
 # Mapping of numpy ndarray dtypes to (field type, subtype)
@@ -611,6 +640,11 @@ cdef detect_encoding(OGRDataSourceH ogr_dataset, OGRLayerH ogr_layer):
         # In old gdal versions, OLCStringsAsUTF8 wasn't advertised yet.
         return "UTF-8"
 
+    if driver == "SQLite":
+        # TestCapability for OLCStringsAsUTF8 returns False for SQLite in GDAL 3.11.3.
+        # Issue opened: https://github.com/OSGeo/gdal/issues/12962
+        return "UTF-8"
+
     return locale.getpreferredencoding()
 
 
@@ -628,8 +662,8 @@ cdef get_fields(OGRLayerH ogr_layer, str encoding, use_arrow=False):
 
     Returns
     -------
-    ndarray(n, 4)
-        array of index, ogr type, name, numpy type
+    ndarray(n, 5)
+        array of index, ogr type, name, numpy type, ogr subtype
     """
     cdef int i
     cdef int field_count
@@ -649,7 +683,7 @@ cdef get_fields(OGRLayerH ogr_layer, str encoding, use_arrow=False):
 
     field_count = OGR_FD_GetFieldCount(ogr_featuredef)
 
-    fields = np.empty(shape=(field_count, 4), dtype=object)
+    fields = np.empty(shape=(field_count, 5), dtype=object)
     fields_view = fields[:, :]
 
     skipped_fields = False
@@ -686,6 +720,7 @@ cdef get_fields(OGRLayerH ogr_layer, str encoding, use_arrow=False):
         fields_view[i, 1] = field_type
         fields_view[i, 2] = field_name
         fields_view[i, 3] = np_type
+        fields_view[i, 4] = field_subtype
 
     if skipped_fields:
         # filter out skipped fields
@@ -880,6 +915,10 @@ cdef process_fields(
     cdef int success
     cdef int field_index
     cdef int ret_length
+    cdef int *ints_c
+    cdef int64_t *int64s_c
+    cdef double *doubles_c
+    cdef char **strings_c
     cdef GByte *bin_value
     cdef int year = 0
     cdef int month = 0
@@ -970,6 +1009,55 @@ cdef process_fields(
                         year, month, day, hour, minute, second, microsecond
                     ).isoformat()
 
+        elif field_type == OFTIntegerList:
+            # According to GDAL doc, this can return NULL for an empty list, which is a
+            # valid result. So don't use check_pointer as it would throw an exception.
+            ints_c = OGR_F_GetFieldAsIntegerList(ogr_feature, field_index, &ret_length)
+            int_arr = np.ndarray(shape=(ret_length,), dtype=np.int32)
+            for j in range(ret_length):
+                int_arr[j] = ints_c[j]
+            data[i] = int_arr
+
+        elif field_type == OFTInteger64List:
+            # According to GDAL doc, this can return NULL for an empty list, which is a
+            # valid result. So don't use check_pointer as it would throw an exception.
+            int64s_c = OGR_F_GetFieldAsInteger64List(
+                ogr_feature, field_index, &ret_length
+            )
+
+            int_arr = np.ndarray(shape=(ret_length,), dtype=np.int64)
+            for j in range(ret_length):
+                int_arr[j] = int64s_c[j]
+            data[i] = int_arr
+
+        elif field_type == OFTRealList:
+            # According to GDAL doc, this can return NULL for an empty list, which is a
+            # valid result. So don't use check_pointer as it would throw an exception.
+            doubles_c = OGR_F_GetFieldAsDoubleList(
+                ogr_feature, field_index, &ret_length
+            )
+
+            double_arr = np.ndarray(shape=(ret_length,), dtype=np.float64)
+            for j in range(ret_length):
+                double_arr[j] = doubles_c[j]
+            data[i] = double_arr
+
+        elif field_type == OFTStringList:
+            # According to GDAL doc, this can return NULL for an empty list, which is a
+            # valid result. So don't use check_pointer as it would throw an exception.
+            strings_c = OGR_F_GetFieldAsStringList(ogr_feature, field_index)
+
+            string_list_index = 0
+            vals = []
+            if strings_c != NULL:
+                # According to GDAL doc, the list is terminated by a NULL pointer.
+                while strings_c[string_list_index] != NULL:
+                    val = strings_c[string_list_index]
+                    vals.append(get_string(val, encoding=encoding))
+                    string_list_index += 1
+
+            data[i] = np.array(vals)
+
 
 @cython.boundscheck(False)  # Deactivate bounds checking
 @cython.wraparound(False)   # Deactivate negative indexing.
@@ -1013,16 +1101,16 @@ cdef get_features(
     field_indexes = fields[:, 0]
     field_ogr_types = fields[:, 1]
 
-    field_data = [
-        np.empty(
-            shape=(num_features, ),
-            dtype = (
-                "object"
-                if datetime_as_string and fields[field_index, 3].startswith("datetime")
-                else fields[field_index, 3]
-            )
-        ) for field_index in range(n_fields)
-    ]
+    field_data = []
+    for field_index in range(n_fields):
+        if datetime_as_string and fields[field_index, 3].startswith("datetime"):
+            dtype = "object"
+        elif fields[field_index, 3].startswith("list"):
+            dtype = "object"
+        else:
+            dtype = fields[field_index, 3]
+
+        field_data.append(np.empty(shape=(num_features, ), dtype=dtype))
 
     field_data_view = [field_data[field_index][:] for field_index in range(n_fields)]
 
@@ -1414,11 +1502,18 @@ def ogr_read(
                 datetime_as_string=datetime_as_string
             )
 
+        ogr_types = [FIELD_TYPE_NAMES.get(field[1], "Unknown") for field in fields]
+        ogr_subtypes = [
+            FIELD_SUBTYPE_NAMES.get(field[4], "Unknown") for field in fields
+        ]
+
         meta = {
             "crs": crs,
             "encoding": encoding,
             "fields": fields[:, 2],
             "dtypes": fields[:, 3],
+            "ogr_types": ogr_types,
+            "ogr_subtypes": ogr_subtypes,
             "geometry_type": geometry_type,
         }
 
@@ -1746,10 +1841,18 @@ def ogr_open_arrow(
         else:
             reader = _ArrowStream(capsule)
 
+        ogr_types = [FIELD_TYPE_NAMES.get(field[1], "Unknown") for field in fields]
+        ogr_subtypes = [
+            FIELD_SUBTYPE_NAMES.get(field[4], "Unknown") for field in fields
+        ]
+
         meta = {
             "crs": crs,
             "encoding": encoding,
             "fields": fields[:, 2],
+            "dtypes": fields[:, 3],
+            "ogr_types": ogr_types,
+            "ogr_subtypes": ogr_subtypes,
             "geometry_type": geometry_type,
             "geometry_name": geometry_name,
             "fid_column": fid_column,
@@ -1906,6 +2009,10 @@ def ogr_read_info(
             encoding = encoding or detect_encoding(ogr_dataset, ogr_layer)
 
         fields = get_fields(ogr_layer, encoding)
+        ogr_types = [FIELD_TYPE_NAMES.get(field[1], "Unknown") for field in fields]
+        ogr_subtypes = [
+            FIELD_SUBTYPE_NAMES.get(field[4], "Unknown") for field in fields
+        ]
 
         meta = {
             "layer_name": get_string(OGR_L_GetName(ogr_layer)),
@@ -1913,6 +2020,8 @@ def ogr_read_info(
             "encoding": encoding,
             "fields": fields[:, 2],
             "dtypes": fields[:, 3],
+            "ogr_types": ogr_types,
+            "ogr_subtypes": ogr_subtypes,
             "fid_column": get_string(OGR_L_GetFIDColumn(ogr_layer)),
             "geometry_name": get_string(OGR_L_GetGeometryColumn(ogr_layer)),
             "geometry_type": get_geometry_type(ogr_layer),