+++ CHANGES.md
@@ -2,9 +2,20 @@
 
 ## 0.12.0 (yyyy-mm-dd)
 
+### Potentially breaking changes
+
+-   Return JSON fields (as identified by GDAL) as dicts/lists in `read_dataframe`;
+    these were previously returned as strings (#556).
+
+### Improvements
+
+-   Add listing of GDAL data types and subtypes to `read_info` (#556).
+-   Add support to read list fields without arrow (#558).
+
 ### Bug fixes
 
--   Fix wrong layername when creating .gpkg.zip file (#570)
+-   Fix decode error reading an sqlite file on windows (#568).
+-   Fix wrong layername when creating .gpkg.zip file (#570).
 
 ## 0.11.1 (2025-08-02)